API: Version viewing on DataObjects
When silverstripe-data-object-version-viewer is installed, adds ability
to view history in data objects to ElementContent, and BaseElement.
Adds a check to BaseElement to determine if history fields should be
added.

WARNING: because this changes field names (Root.Main becomes
Root.Current.Main), there could be some integration required on upgrade
if you have used extensions with updateCMSFields on ElementContent, or
BaseElement .