NEW add ElementalArea field type
Which subclasses GridField and behaves with no altered functionality
but upon a read only transformation renders out it's associated Elements
(the items in the grid field). By Render out here this means that a
summary of sorts by the way of the 'Root.Main' tab of said element's
getCMSFields.

This in turns allows a page to feature history complete with a content
summary when inspecting a version, since this module does away with
the default Content field.