Add unit tests for new page level history view
During the course of the last few commits functionality has been added
that allows a user to view a full collection of data for an
ElementalArea at a Page level. This has been implemented via the
introduction of a new FormField type, which until now has not been
tested in either the front (react) end or the back (php) end.