Alter elemenal history to conditionally transform
The previous iteration relied on a core (silverstripe/admin) alteration
to add extra context to permit the transformation via context alone.
Instead of relying on this we can conditionally tranform within the
higher order component, based on the presence or abescence of properties
it requires to do it's thing (i.e. ElementID).