FIX push element history revert button to bottom of page
Previously both the revert button shows immediately under the content,
and the save/publish buttons show at the bottom of the page (in the CMS)
as per the usual actions for an edit form (for which the
HistoryViewerField is added to under a tab of it's own). Clearly
confusing, a work around has been introduced to
silverstripe/versioned-admin in order work around this - this commit
enables that functionality which has been made opt-in as the nature of
it being a work-around.