Add ability to load Element's edit form inline
For allowing a user to edit that element and save the changes without
having to navigate through multiple page loads. Uses form schema for this
puspose, and relies on correlating changes in silverstripe/admin and
silverstripe/framework in order to render the one bundled Element -
Content (which requires info from the PHP side in order to render
TinyMCE).