FIX add-element-hover-bar layout issues in IE
Internet Explorer 11 and MS Edge browser each have their own
idiosyncracies when it comes to layout (...and Safari too).
To get things to a visulally similar state between all major browsers
a bit of rearragning the (S)CSS needed to take place. Things have been
fixed such as the hover bar not rendering at the base of an element, and
the plus icon rendering extremely off-centre from the hover bar.