Connect mutation to add blocks to relevant UI elements
As part of inline editing for Elemental it has previously been only
possible to edit existing elements. Now it is possible to add new
ones to the relevant list for the pop over - both the add button
above the list, or the in-between button that pops up on hovering
the area close to the junction between two elements in the list.

For this work it was necessary to expose the ID of the elemental area
they belong to, in order to ensure elements are added to it as children.

There have been a few scss linting issues tidied up also.