NEW Give user better feedback on adding elements
Previously when adding an element to an elemental area it would
default to having no title, and no content (of any type). They now
are given some filler text for a title, at a different shade to
indicate the fact that it is not a 'real' title, and all elements
have a min-height... not by explictly setting this, but by always
drawing the content area summary paragraph (even if it's empty).

This also solves the case of when there is only one element in the
area, one could see the area 'sticking out' from behind the element.
Because the element is now larger (mostly padding and margins) this
no longer happens.