NEW Don't render InlineEditForm on loadingError. Make ElementalAreaController inherit from CMSMain
Implement an onLoadingError method to be passed to FormBuilderLoader. This let's us know in InlineEditForm is an error has occured and take necessary action, such as preventing the Element component from expanding the Content. When a loading error occurs the InlineLineEditForm cannot be rendered, this would cause the Summary component to be toggled erroneously.
It makes more sense for ElementalAreaController to inherit from CMSMain rather than LeftAndMain, as this means that permissions are checked in relation to the Page Level rather than for the entire CMS.