FIX update tab loading to use recent admin module changes
The admin module previously used the redux-form store to track the
active Tab within a Tabs component. However this was recently refactored
in the admin module to use a different shard in the redux store, as it
was causing problems (functionality regressions) in form state with the
asset-admin module. So the logic here in Elemental also required updating.