Update form styles in CMS when editing Elements
The fields in e.g. 'split mode' editing within the CMS looked a bit
squished as the labels first took up an amount of room to the left of
the field proper. The better and desired approach here was to use the
.stacked class - rendering the labels atop the field.

However this approach would require every element built to add the
extraClass to the FormField instance in PHP in the getCMSFields method.
This is less than ideal, so it was decided to proceed with an override
for elements (and elements only) that will achieve the same effect.

Also addresses checkboxes and radio buttons where the label generally
sits on the other side of the input element, where .stacked does not.