FIX always build an ElementalArea for all newly extended classes
Previously only brand new pages would have ElementalAreas built for
them, as this process happens on write. If an existing page type was
updated to have ElementalAreas on it, all existing pages in the CMS
would error when trying to be viewed, as the ElementalArea would not
exist. This was particularly true for GraphQL queries, which are
abundant in the upcoming 4.0 release.