Fix incorrect ElementalArea.OwnerClassName on save
OwnerClassName in the ElementalArea table can become incorrect in certain scenarios*, and there is currently no way to fix this short of manually editing the database.  An incorrect OwnerClassName breaks the behaviour of BaseElement->getPage(), which breaks all kinds of other things in BaseElement that depend on it.

For example BaseElement->canView() will always return false.  I suspect a lot of people are working around this issue by putting canView{ return true } in all their block types.  That's what we were doing until I decided to really dig into the root cause of why BaseElement->canView() was failing.

This change modifies ElementalAreasExtension->ensureElementalAreasExist(), to verify that the OwnerClassName is correct, and re-saves it with the correct value if it is wrong.

*For example, apply ElementalPageExtension to Page, create a Page, then switch ElementalPageExtension to OtherPage, and switch your created page type to OtherPage.  The OwnerClassName in the DB will still be "Page" and there is no way to fix this currently.