NEW ability to configure content field replacement
Globally or per class the ElementalAreasExtension is applied to.
This allows for flexibility where a developer may want Elemental to
detour from its default functionality that removes the Content field
on whatever object it is applied to (for whatever reason, e.g. migration
from another module that did not replace the field).

This change defines that settings applied to the class the extension is applied to
will override the global setting. This allows for both new page types to retain
elemental's default functionality, while providing the flexibility to retain e.g.
migrated content on older page types.