NEW Introduce error boundary to Element for softer crashes
Most exceptions in the ElementalEditor are caused by a malfunction in an
element's form components. Prior to this change, React would unmount the
entire editor when an exception like this was thrown, resulting in
confusion for users. Now, we capture these exceptions at the Element
component level, and render a warning, instructing the user to attempt
recovery by saving and reloading.