DEP Restore GraphQL 4 support
This was originally added to the 4 release branch,
and then removed prior to the 4.5.0 release through
https://github.com/dnadesign/silverstripe-elemental/pull/851
to simplify Travis builds for the release activity.

Explicitly declaring the module *doesn't* support GraphQL 4
is incorrect, and hinders early adoption - you'll need to
run a "silverstripe/graphql:4.x-dev as 3.x-dev" dependency.

It's also a different strategy compared to e.g. versioned-admin:
https://github.com/silverstripe/silverstripe-versioned-admin/commit/cae464fd25494e24d1e05d863c7bc1d4221d8d57

Relies on https://github.com/silverstripe/silverstripe-travis-shared/pull/4