Prevent mapStateToProps re-rendering on formstate changes
any formstate change, eg from another FormLoader component will
trigger the ElementalEditor to re-render because this function
creates a new object for every run. This change assumes form state
does not change after it is created. If it needs to, it could be
that we need to use another way to pass down that difference.