API Update search functionality to work in a SearchContext
This relied on the weird non-standard way CMSMain page filtering worked.
This same functionality could be reintroduced by replacing the
SiteTreeSearchContext with a custom one that filters by callback, but it
won't be available in this module.

Some improvements have also been made to avoid filtering through pages
that don't have the relevant extension, and which are explicitly set to
be ignored by the extension.